hotfix: fix marks for second review

+++ backend/controllers/cards.js
@@ -22,20 +22,19 @@ const createCard = (req, res, next) => {
 const deleteCard = (req, res, next) => {
   Card.findById(req.params._id)
     .then((result) => {
+      if (!result) {
+        throw new NotFoundError('Карточка с таким id не найдена');
+      }
       /* eslint-disable eqeqeq */
       if (req.user._id == result.owner) {
       /* eslint-enable eqeqeq */
         Card.findByIdAndRemove(req.params._id)
           .then((card) => {
-            if (!card) {
-              throw new NotFoundError('Карточка с таким id не найдена');
-            }
-            return res.send(card);
+            res.send(card);
           })
           .catch(next);
       } else {
-        const err = new AuthError('Невозможно удалить чужую карточку');
-        next(err);
+        throw new AuthError('Невозможно удалить чужую карточку');
       }
     })
     .catch(next);
